Clean and add log, duplicate check hook. Feature: - Add log submodule for `flow`. - Add check hook for AbacusFlow.(Show state status and task list. See `check_exist_status` of calculation_state) - Add default value of nbands and kpointfix. - Add null value test before writing abacus INPUT file. - Add null value test before submission generate. - Add null value test before run_submission.

+++ src/bandapi/flow/abacus/calculation_state.py
@@ -7,22 +7,19 @@
 # ====================================== #
 import glob
 import pathlib
+import re
 import shutil
 from collections import Counter
 
 import ase
 import ase.symbols
 import numpy as np
-from ase.atoms import Atoms
 from bandapi.dispatcher.dpdispatcher import Task
 from bandapi.flow.abacus import default_settings
-from bandapi.flow.abacus.utils import write_stur
-from bandapi.flow.state import FlowState, FlowStateControl
+from bandapi.flow.state import FlowStateControl
 from bandapi.flow.task_content import NamedAtomsContentDict
-from bandapi.io.abacus.input import write_abacus_input
-from bandapi.io.abacus.kpt import write_abacus_kpt
-from bandapi.io.abacus.potential import AbacusPotential
 from bandapi.io.abacus.out import read_stru
+from bandapi.io.abacus.potential import AbacusPotential
 
 """
 Abacus has calculation state as following:
@@ -46,22 +43,29 @@ class AbacusScfState(AbacusState):
         :return:
         """
         for subdir, atoms in task_content.items():
-            atoms: ase.Atoms
-            self._write_stur(subname=subdir, atoms=atoms, potential_name=self.get_state_settings("potential_name"))
-            self._write_kpt(subname=subdir, atoms=atoms)
-            self._write_input(subname=subdir, atoms=atoms)
+            if hasattr(self, "check_exist_status"):
+                self.check_exist_status: dict
+                if not self.check_exist_status.get(subdir, None):
+                    atoms: ase.Atoms
+                    self._write_stur(subname=subdir, atoms=atoms, potential_name=self.get_state_settings("potential_name"))
+                    self._write_kpt(subname=subdir, atoms=atoms)
+                    self._write_input(subname=subdir, atoms=atoms)
 
     def prepare(self, task_content: NamedAtomsContentDict, task_settings):
         task_list = []
         for idx, item in enumerate(task_content.keys()):
             for idx, item in enumerate(task_content):
-                task_list.append(
-                    Task(command=task_settings["remote_command"],
-                         task_work_path=f"{self._state}/{item}/",
-                         forward_files=[*self.bake_upload_files(task_content[item])],
-                         backward_files=["OUT.ABACUS"]
-                         )
-                )
+                if hasattr(self, "check_exist_status"):
+                    self.check_exist_status: dict
+                    if not self.check_exist_status.get(item, None):
+                        task_list.append(
+                            Task(command=task_settings["remote_command"],
+                                 task_work_path=f"{self._state}/{item}/",
+                                 forward_files=[*self.bake_upload_files(task_content[item])],
+                                 backward_files=["OUT.ABACUS"]
+                                 ))
+                    else:
+                        pass
         return task_list
 
     def run_end(self, next_state: str):
@@ -104,13 +108,22 @@ class AbacusScfState(AbacusState):
         :param ase.Atoms atoms:
         :return:
         """
-        scope = self.get_state_settings("kpointscope", default_settings["kpointscope"])
-        odd_flag = scope % 2
-        abc = atoms.cell.lengths()
-        result = np.around(1 / abc / min(1 / abc) * scope)
-        mask = result % 2 != odd_flag
-        shift = np.zeros_like(result)
-        content = np.concatenate([result + mask, shift], axis=-1)
+        if not self.get_state_settings("kpointfix", default_settings["kpointfix"]):
+            scope = self.get_state_settings("kpointscope", default_settings["kpointscope"])
+            odd_flag = scope % 2
+            abc = atoms.cell.lengths()
+            result = np.around(1 / abc / min(1 / abc) * scope)
+            if result[0] * result[1] * result[2] > 1000:
+                scope -= 2
+                odd_flag = scope % 2
+                abc = atoms.cell.lengths()
+                result = np.around(1 / abc / min(1 / abc) * scope)
+            mask = result % 2 != odd_flag
+            shift = np.zeros_like(result)
+            content = np.concatenate([result + mask, shift], axis=-1)
+        else:
+            scope = self.get_state_settings("kpointscope", default_settings["kpointscope"])
+            content = np.array([scope, scope, scope, 0, 0, 0])
         return {
             "number_of_kpt": 0,
             "mode": "Gamma",
@@ -153,7 +166,7 @@ class AbacusRelaxState(AbacusScfState, AbacusState):
             "symmetry": 0,
             "ecutwfc": self.get_state_settings("ecutwfc", default_settings["ecutwfc"]),
             "dr2": self.get_state_settings("dr2", default_settings["dr2"]),
-            "nstep":self.get_state_settings("nstep", default_settings["nstep"]),
+            "nstep": self.get_state_settings("nstep", default_settings["nstep"]),
         }
 
     def run_end(self, next_state: str):
@@ -166,9 +179,9 @@ class AbacusRelaxState(AbacusScfState, AbacusState):
         if next_state is None:
             pass
         elif next_state is not None:
-            self._submit_loop_condition=1
+            self._submit_loop_condition = 1
             for subdir, _ in self.task_content.items():
-                self.task_content[subdir]=read_stru(self.flow_work_root/self._state/subdir/"OUT.ABACUS"/"STRU_ION_D")
+                self.task_content[subdir] = read_stru(self.flow_work_root / self._state / subdir / "OUT.ABACUS" / "STRU_ION_D")
             return True
         else:
             raise NotImplementedError
@@ -194,7 +207,7 @@ class AbacusCellRelaxState(AbacusScfState, AbacusState):
             "symmetry": 0,
             "ecutwfc": self.get_state_settings("ecutwfc", default_settings["ecutwfc"]),
             "dr2": self.get_state_settings("dr2", default_settings["dr2"]),
-            "nstep":self.get_state_settings("nstep", default_settings["nstep"]),
+            "nstep": self.get_state_settings("nstep", default_settings["nstep"]),
         }
 
     def run_end(self, next_state: str):
@@ -206,10 +219,10 @@ class AbacusCellRelaxState(AbacusScfState, AbacusState):
         """
         if next_state is None:
             pass
-        elif next_state == "scf"or"relax"or"scf-charge":
+        elif next_state == "scf" or "relax" or "scf-charge":
             self._submit_loop_condition = 1
             for subdir, _ in self.task_content.items():
-                self.task_content[subdir]=read_stru(self.flow_work_root/self._state/subdir/"OUT.ABACUS"/"STRU_ION_D")
+                self.task_content[subdir] = read_stru(self.flow_work_root / self._state / subdir / "OUT.ABACUS" / "STRU_ION_D")
             return True
         else:
             raise NotImplementedError
@@ -217,6 +230,16 @@ class AbacusCellRelaxState(AbacusScfState, AbacusState):
 
 class AbacusScfStateWithCharge(AbacusScfState, AbacusState):
     _state = "scf-charge"
+
+    def flow_begin_test(self):
+        check_status = {}
+        for subdir, atoms in self.task_content.items():
+            atoms: ase.Atoms
+            CHGfile = glob.glob((self.flow_work_root / "scf-charge" / subdir / "OUT.ABACUS" / "SPIN*_CHG").as_posix())
+            for item in CHGfile:
+                if item:
+                    check_status[subdir] = True
+        self.check_exist_status = check_status
 
     def get_input_args(self, atoms):
         """
@@ -250,7 +273,7 @@ class AbacusScfStateWithCharge(AbacusScfState, AbacusState):
         elif next_state == "nscf-band":
             self._submit_loop_condition = 1
             for subdir, _ in self.task_content.items():
-                self.task_content[subdir]=read_stru(self.flow_work_root/self._state/subdir/"STRU")
+                self.task_content[subdir] = read_stru(self.flow_work_root / self._state / subdir / "STRU")
             return True
         else:
             raise NotImplementedError
@@ -260,12 +283,21 @@ class AbacusBandState(AbacusState):
     _state = "nscf-band"
 
     def flow_begin_test(self):
+        check_status = {}
         for subdir, atoms in self.task_content.items():
             atoms: ase.Atoms
-            CHGfile = glob.glob((self.flow_work_root/"scf-charge"/subdir/"OUT.ABACUS"/"SPIN*_CHG").as_posix())
-            (self.flow_work_root / self._state / subdir / "OUT.ABACUS").mkdir(parents=True,exist_ok=True)
+            CHGfile = glob.glob((self.flow_work_root / self._state / subdir / "OUT.ABACUS" / "running_nscf*").as_posix())
             for item in CHGfile:
-                shutil.copy(item,self.flow_work_root/self._state/subdir/"OUT.ABACUS/")
+                if item:
+                    check_status[subdir] = True
+        self.check_exist_status = check_status
+        for subdir, atoms in self.task_content.items():
+            if not self.check_exist_status.get(subdir,None):
+                atoms: ase.Atoms
+                CHGfile = glob.glob((self.flow_work_root / "scf-charge" / subdir / "OUT.ABACUS" / "SPIN*_CHG").as_posix())
+                (self.flow_work_root / self._state / subdir / "OUT.ABACUS").mkdir(parents=True, exist_ok=True)
+                for item in CHGfile:
+                    shutil.copy(item, self.flow_work_root / self._state / subdir / "OUT.ABACUS/")
 
     def bakeup(self, task_content: NamedAtomsContentDict):
         """
@@ -274,21 +306,28 @@ class AbacusBandState(AbacusState):
         :return:
         """
         for subdir, atoms in task_content.items():
-            self._write_stur(subname=subdir, atoms=atoms, potential_name=self.get_state_settings("potential_name"))
-            self._write_kpt(subname=subdir, atoms=atoms)
-            self._write_input(subname=subdir, atoms=atoms)
+            if hasattr(self, "check_exist_status"):
+                self.check_exist_status: dict
+                if not self.check_exist_status.get(subdir, None):
+                    for subdir, atoms in task_content.items():
+                        self._write_stur(subname=subdir, atoms=atoms, potential_name=self.get_state_settings("potential_name"))
+                        self._write_kpt(subname=subdir, atoms=atoms)
+                        self._write_input(subname=subdir, atoms=atoms)
 
     def prepare(self, task_content: NamedAtomsContentDict, task_settings):
         task_list = []
         for idx, item in enumerate(task_content.keys()):
             for idx, item in enumerate(task_content):
-                task_list.append(
-                    Task(command=task_settings["remote_command"],
-                         task_work_path=f"{self._state}/{item}/",
-                         forward_files=[*self.bake_upload_files(task_content[item])],
-                         backward_files=["OUT.ABACUS"]
-                         )
-                )
+                if hasattr(self, "check_exist_status"):
+                    self.check_exist_status: dict
+                    if not self.check_exist_status.get(item, None):
+                        task_list.append(
+                            Task(command=task_settings["remote_command"],
+                                 task_work_path=f"{self._state}/{item}/",
+                                 forward_files=[*self.bake_upload_files(task_content[item])],
+                                 backward_files=["OUT.ABACUS"]
+                                 )
+                        )
         return task_list
 
     def run_end(self, next_state: str):
@@ -300,7 +339,7 @@ class AbacusBandState(AbacusState):
         """
         if next_state is None:
             return False
-        elif next_state=="band-data":
+        elif next_state == "band-data":
             return True
         else:
             return NotImplementedError
@@ -317,6 +356,7 @@ class AbacusBandState(AbacusState):
         return {
             "pseudo_dir": "./",
             "calculation": "nscf",
+            "nbands": self.get_state_settings("nbands", default_settings["nbands"]),
             "ntype": len(atom_type_list),
             "basis_type": "pw",
             "symmetry": 0,
@@ -339,7 +379,7 @@ class AbacusBandState(AbacusState):
         path_lines = []
         num_lines = []
         for item in path:
-            for point in item:
+            for point in re.findall("\w\d*", item):
                 path_lines.append(sp[point])
                 num_lines.append(scope)
